+++ lib/fstab/zfs.rb
@@ -1,4 +1,7 @@
+# frozen_string_literal: true
+
 module Fstab
+  # Generating fstab for zfs filesystem
   class Zfs < Root
     def initialize(devs, options)
       super
@@ -14,9 +17,11 @@ module Fstab
 
     def write_swap
       uuid = gen_uuid @swap
-      @encrypt ?
-        line = "/dev/mapper/swap-luks none swap sw 0 0" :
-        line = "UUID=#{uuid} swap swap rw,noatime,discard 0 0"
+      line = if @encrypt
+               '/dev/mapper/swap-luks none swap sw 0 0'
+             else
+               "UUID=#{uuid} swap swap rw,noatime,discard 0 0"
+             end
       echo_a @conf, line
     end
   end