+++ lib/getch/void/pre_config.rb
@@ -2,6 +2,7 @@
 
 module Getch
   module Void
+    # config before installing the system
     class PreConfig
       def initialize
         x